pyproject.toml
[project]
name = "odds-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "mysql-connector-python>=9.2.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "tqdm>=4.67.1",
]
